[joblib_launcher] cast params to int when applicable (#1098) (#1114)
* Cast params to int when applicable

+++ plugins/hydra_joblib_launcher/hydra_plugins/hydra_joblib_launcher/config.py
@@ -24,8 +24,8 @@ class JobLibLauncherConf:
     # if greater than zero, prints progress messages
     verbose: int = 0
 
-    # timeout limit for each task
-    timeout: Optional[int] = None
+    # timeout limit for each task. Unit dependent on backend implementation; miliseconds for loky.
+    timeout: Optional[float] = None
 
     # number of batches to be pre-dispatched
     pre_dispatch: str = "2*n_jobs"
